+++ src/js/app.js
@@ -1,6 +1,6 @@
 import LocalistView from 'localist-viewer';
 
-// const data = require('./config.json');
+const data = require('./config.json');
 require('babel-polyfill');
 
 /** *********************** START LOCALIST *********************************** /
@@ -10,13 +10,15 @@ require('babel-polyfill');
  *  self invoking anonymous function, using drupal behaviours API.
  *
  **************************************************************************** */
-// LocalistView(data);
+LocalistView(data);
 
-// another example of how to config
-const eventListings = [
-    ...document.getElementsByClassName('events-listing')
-];
-eventListings.forEach(elem => {
-    const data = { ...elem.dataset };
-    LocalistView(data);
-});
+console.log(data);
+
+// // another example of how to config
+// const eventListings = [
+//     ...document.getElementsByClassName('events-listing')
+// ];
+// eventListings.forEach(elem => {
+//     const data = { ...elem.dataset };
+//     LocalistView(data);
+// });